+++ lib/in_app_update.dart
@@ -2,6 +2,28 @@ import 'dart:async';
 
 import 'package:flutter/services.dart';
 
+class InstallStatus {
+  InstallStatus._();
+
+  static int get unknown => 0;
+  static int get pending => 1;
+  static int get downloading => 2;
+  static int get installing => 3;
+  static int get installed => 4;
+  static int get failed => 5;
+  static int get canceled => 6;
+  static int get downloaded => 11;
+}
+
+class UpdateAvailability {
+  UpdateAvailability._();
+
+  static int get unknown => 0;
+  static int get updateNotAvailable => 1;
+  static int get updateAvailable => 2;
+  static int get developerTriggeredUpdateInProgress => 3;
+}
+
 class InAppUpdate {
   static const MethodChannel _channel = const MethodChannel('in_app_update');
 
@@ -11,8 +33,18 @@ class InAppUpdate {
   /// [startFlexibleUpdate] or [performImmediateUpdate] should be called.
   static Future<AppUpdateInfo> checkForUpdate() async {
     final result = await _channel.invokeMethod('checkForUpdate');
-    return AppUpdateInfo(result['updateAvailable'], result['immediateAllowed'], result['flexibleAllowed'],
-        result['availableVersionCode']);
+
+    return AppUpdateInfo(
+      result['updateAvailability'] == UpdateAvailability.updateAvailable,
+      result['updateAvailability'],
+      result['immediateAllowed'],
+      result['flexibleAllowed'],
+      result['availableVersionCode'],
+      result['installStatus'],
+      result['packageName'],
+      result['clientVersionStalenessDays'],
+      result['updatePriority'],
+    );
   }
 
   /// Performs an immediate update that is entirely handled by the Play API.
@@ -41,32 +73,60 @@ class InAppUpdate {
 }
 
 class AppUpdateInfo {
-  final bool updateAvailable, immediateUpdateAllowed, flexibleUpdateAllowed;
-  final int availableVersionCode;
+  @Deprecated('Use updateAvailability instead.')
+  final bool updateAvailable;
+  final bool immediateUpdateAllowed, flexibleUpdateAllowed;
+  final int updateAvailability,
+      availableVersionCode,
+      installStatus,
+      clientVersionStalenessDays,
+      updatePriority;
+  final String packageName;
 
   AppUpdateInfo(
-      this.updateAvailable, this.immediateUpdateAllowed, this.flexibleUpdateAllowed, this.availableVersionCode);
+    this.updateAvailable,
+    this.updateAvailability,
+    this.immediateUpdateAllowed,
+    this.flexibleUpdateAllowed,
+    this.availableVersionCode,
+    this.installStatus,
+    this.packageName,
+    this.clientVersionStalenessDays,
+    this.updatePriority,
+  );
 
   @override
   bool operator ==(Object other) =>
       identical(this, other) ||
       other is AppUpdateInfo &&
           runtimeType == other.runtimeType &&
-          updateAvailable == other.updateAvailable &&
+          updateAvailability == other.updateAvailability &&
           immediateUpdateAllowed == other.immediateUpdateAllowed &&
           flexibleUpdateAllowed == other.flexibleUpdateAllowed &&
-          availableVersionCode == other.availableVersionCode;
+          availableVersionCode == other.availableVersionCode &&
+          installStatus == other.installStatus &&
+          packageName == other.packageName &&
+          clientVersionStalenessDays == other.clientVersionStalenessDays &&
+          updatePriority == other.updatePriority;
 
   @override
   int get hashCode =>
-      updateAvailable.hashCode ^
+      updateAvailability.hashCode ^
       immediateUpdateAllowed.hashCode ^
       flexibleUpdateAllowed.hashCode ^
-      availableVersionCode.hashCode;
+      availableVersionCode.hashCode ^
+      installStatus.hashCode ^
+      packageName.hashCode ^
+      clientVersionStalenessDays.hashCode ^
+      updatePriority.hashCode;
 
   @override
-  String toString() => 'InAppUpdateState{updateAvailable: $updateAvailable, '
+  String toString() => 'InAppUpdateState{updateAvailability: $updateAvailability, '
       'immediateUpdateAllowed: $immediateUpdateAllowed, '
       'flexibleUpdateAllowed: $flexibleUpdateAllowed, '
-      'availableVersionCode: $availableVersionCode}';
+      'availableVersionCode: $availableVersionCode, '
+      'installStatus: $installStatus, '
+      'packageName: $packageName, '
+      'clientVersionStalenessDays: $clientVersionStalenessDays, '
+      'updatePriority: $updatePriority}';
 }